feat: add crud to existing models and add authorization in backend

+++ server/WebApi/Controllers/MeterController.cs
@@ -1,12 +1,15 @@
-using Application.DTOs;
+using System.Security.Claims;
 using Application.Interfaces;
 using Domain.Exceptions;
+using Domain.Models;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace WebApi.Controllers;
 
+[Authorize(Roles = "User")]
 [ApiController]
-[Route("[controller]")]
+[Route("api/[controller]")]
 public class MeterController : ControllerBase
 {
     private readonly ILogger<MeterController> _logger;
@@ -19,15 +22,19 @@ public class MeterController : ControllerBase
     }
 
     [HttpGet]
-    [ProducesResponseType(typeof(MeterDto[]), StatusCodes.Status200OK)]
+    [ProducesResponseType(typeof(IEnumerable<Meter>), StatusCodes.Status200OK)]
     public async Task<IActionResult> GetAll()
     {
-        var meters = await _meterService.GetAll();
+        var userId = int.Parse(User.FindFirst(ClaimTypes.NameIdentifier)?.Value ?? "-1");
+        if (userId == -1)
+            return Unauthorized();
+
+        var meters = await _meterService.GetByUserId(userId);
         return Ok(meters);
     }
 
     [HttpGet("{id}")]
-    [ProducesResponseType(typeof(MeterDto), StatusCodes.Status200OK)]
+    [ProducesResponseType(typeof(Meter), StatusCodes.Status200OK)]
     [ProducesResponseType(StatusCodes.Status404NotFound)]
     [ProducesResponseType(StatusCodes.Status500InternalServerError)]
     public async Task<IActionResult> Get(int id)
@@ -35,6 +42,13 @@ public class MeterController : ControllerBase
         try
         {
             var meter = await _meterService.GetById(id);
+
+            if (
+                meter.UserId != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier)?.Value ?? "-1")
+                && !User.FindAll(ClaimTypes.Role).Any(x => x?.Value == "Admin")
+            )
+                return Unauthorized();
+
             return Ok(meter);
         }
         catch (Exception ex)
@@ -48,9 +62,15 @@ public class MeterController : ControllerBase
     }
 
     [HttpPost]
-    [ProducesResponseType(typeof(MeterDto), StatusCodes.Status201Created)]
-    public async Task<IActionResult> Create([FromBody] MeterDto meter)
+    [ProducesResponseType(typeof(Meter), StatusCodes.Status201Created)]
+    public async Task<IActionResult> Create([FromBody] Meter meter)
     {
+        if (
+            meter.UserId != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier)?.Value ?? "-1")
+            && !User.FindAll(ClaimTypes.Role).Any(x => x?.Value == "Admin")
+        )
+            return Unauthorized();
+
         var createdMeter = await _meterService.Create(meter);
         return CreatedAtAction(nameof(Create), new { id = createdMeter.Id }, createdMeter);
     }
@@ -60,10 +80,16 @@ public class MeterController : ControllerBase
     [ProducesResponseType(StatusCodes.Status404NotFound)]
     [ProducesResponseType(StatusCodes.Status400BadRequest)]
     [ProducesResponseType(StatusCodes.Status500InternalServerError)]
-    public async Task<IActionResult> Update(int id, [FromBody] MeterDto meter)
+    public async Task<IActionResult> Update(int id, [FromBody] Meter meter)
     {
         if (id != meter.Id)
             return BadRequest("Meter ID mismatch");
+
+        if (
+            meter.UserId != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier)?.Value ?? "-1")
+            && !User.FindAll(ClaimTypes.Role).Any(x => x?.Value == "Admin")
+        )
+            return Unauthorized();
 
         try
         {
@@ -88,6 +114,14 @@ public class MeterController : ControllerBase
     {
         try
         {
+            var meter = await _meterService.GetById(id);
+
+            if (
+                meter.UserId != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier)?.Value ?? "-1")
+                && !User.FindAll(ClaimTypes.Role).Any(x => x?.Value == "Admin")
+            )
+                return Unauthorized();
+
             await _meterService.Delete(id);
             return NoContent();
         }
